requirements.txt
Crypto
requests
lxml
mutagen
pillow

# additional requirements for GUI
tkinter
pyinstaller

